AdAuth/requirements.txt
ldap3==2.9.1
pytest==8.3.2
pytest-mock==3.14.0
pytest-freezer==0.4.8
pyyaml==6.0.2
flake8==7.1.1
black==24.8.0
isort==5.13.2
mypy==1.11.1
types-ldap3==2.9.13
pre-commit==3.8.0
